Add files via upload
I have had some problems with my server if not many. While the Client runs, the server has some issues on my computer. I have not been able to make a game as yet, as I have started out late. However I have tried to understand the code of server and client. So this code is merely foccusing on the communication of the two. I want to reference some of the code from https://www.baeldung.com/a-guide-to-java-sockets and https://www.pegaxchange.com/2017/12/07/simple-tcp-ip-server-client-java/